dbeaver/pro#9017 Move timeout, logging and temperature to base props class
