+++ plugins/org.jkiss.dbeaver.model.ai/src/org/jkiss/dbeaver/model/ai/engine/copilot/CopilotProperties.java
@@ -21,20 +21,17 @@ import org.jkiss.code.NotNull;
 import org.jkiss.code.Nullable;
 import org.jkiss.dbeaver.DBException;
 import org.jkiss.dbeaver.model.ai.AIConfigurationProfile;
-import org.jkiss.dbeaver.model.ai.engine.AIEngineProperties;
 import org.jkiss.dbeaver.model.ai.engine.AIModel;
+import org.jkiss.dbeaver.model.ai.engine.BaseAIEngineProperties;
 import org.jkiss.dbeaver.model.ai.utils.AIUtils;
 import org.jkiss.dbeaver.model.meta.Property;
 import org.jkiss.dbeaver.model.meta.SecureProperty;
 import org.jkiss.utils.CommonUtils;
 
-public class CopilotProperties implements AIEngineProperties {
+public class CopilotProperties extends BaseAIEngineProperties {
     private static final String COPILOT_ACCESS_TOKEN = "copilot.access.token";
     private static final String GPT_MODEL = "gpt.model";
     private static final String GPT_CONTEXT_WINDOW_SIZE = "gpt.contextWindowSize";
-    private static final String GPT_MODEL_TEMPERATURE = "gpt.model.temperature";
-    private static final String GPT_LOG_QUERY = "gpt.log.query";
-    private static final String GPT_TIMEOUT = "gpt.timeout";
 
     @Nullable
     @SecureProperty
@@ -48,16 +45,6 @@ public class CopilotProperties implements AIEngineProperties {
     @Nullable
     @SerializedName(GPT_CONTEXT_WINDOW_SIZE)
     private Integer contextWindowSize;
-
-    @SerializedName(GPT_MODEL_TEMPERATURE)
-    private double temperature;
-
-    @SerializedName(GPT_LOG_QUERY)
-    private boolean loggingEnabled;
-
-    @Nullable
-    @SerializedName(GPT_TIMEOUT)
-    private Integer timeout;
 
     @Nullable
     @Property(order = 1, password = true, required = true)
@@ -95,10 +82,6 @@ public class CopilotProperties implements AIEngineProperties {
             .orElse(AIUtils.DEFAULT_TEMPERATURE);
     }
 
-    public void setTemperature(double temperature) {
-        this.temperature = AIUtils.normalizeTemperature(temperature);
-    }
-
     @Override
     @Nullable
     @Property(order = 4)
@@ -114,26 +97,6 @@ public class CopilotProperties implements AIEngineProperties {
 
     public void setContextWindowSize(@Nullable Integer contextWindowSize) {
         this.contextWindowSize = contextWindowSize;
-    }
-
-    @Property(order = 5)
-    public boolean isLoggingEnabled() {
-        return loggingEnabled;
-    }
-
-    public void setLoggingEnabled(boolean loggingEnabled) {
-        this.loggingEnabled = loggingEnabled;
-    }
-
-    @Override
-    @Property(order = 6)
-    public int getTimeout() {
-        return timeout != null && timeout > 0 ? timeout : AIEngineProperties.DEFAULT_TIMEOUT;
-    }
-
-    @Override
-    public void setTimeout(int timeout) {
-        this.timeout = timeout;
     }
 
     /**
